[DEV] Front-end e Back-end
- FRONT: Altera card de produtos e implementa página de carrinho;
- BACK: Adicionar carrinho controller, service, model e repository;
- BACK: Implementa exclusão de carrinho se for deletar um produto que esteja no carrinho;

+++ back/src/main/java/com/ecommerce/back/service/ProdutoService.java
@@ -1,15 +1,21 @@
 package com.ecommerce.back.service;
 
 import com.ecommerce.back.dto.DadosProduto;
+import com.ecommerce.back.model.Carrinho;
 import com.ecommerce.back.model.Produto;
+import com.ecommerce.back.repository.CarrinhoRepository;
 import com.ecommerce.back.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ProdutoService {
+
+    @Autowired
+    CarrinhoRepository carrinhoRepository;
 
     @Autowired
     ProdutoRepository produtoRepository;
@@ -19,7 +25,12 @@ public class ProdutoService {
     }
 
     public Produto insereProduto(DadosProduto dados) {
-        Produto produto = new Produto(dados);
+        Produto produto = Produto.builder()
+                .nome(dados.nome())
+                .preco(dados.preco())
+                .quantidade(dados.quantidade())
+                .imagem(dados.imagem()).build();
+
         produtoRepository.save(produto);
         return produto;
     }
@@ -35,15 +46,31 @@ public class ProdutoService {
         return produto;
     }
 
-    public void excluirFuncionario(Long id) {
-        produtoRepository.delete(buscaProduto(id));
+    public void excluirProduto(Long id) {
+        if (produtoEstaNoCarrinho(id)) {
+            findProdutoNoCarrinho(id).forEach(produto -> carrinhoRepository.delete(produto));
+            produtoRepository.delete(buscaProduto(id));
+        }
     }
 
     public Produto detalhaFuncionario(Long id) {
         return buscaProduto(id);
     }
 
+
+    private Carrinho buscaCarrinho(Long id) {
+        return carrinhoRepository.findById(id).get();
+    }
+
     private Produto buscaProduto(Long id) {
         return produtoRepository.findById(id).get();
+    }
+
+    private boolean produtoEstaNoCarrinho(Long id) {
+        return !findProdutoNoCarrinho(id).isEmpty();
+    }
+
+    private List<Carrinho> findProdutoNoCarrinho(Long id) {
+        return carrinhoRepository.findAll().stream().filter(carrinho -> carrinho.getProduto().getId().equals(id)).collect(Collectors.toList());
     }
 }
